requirements.txt
python-telegram-bot==21.5
httpx==0.27.0
openai==1.40.0
duckduckgo-search==6.2.6
SQLAlchemy==2.0.31
aiosqlite==0.20.0
python-dotenv==1.0.1
aiofiles==23.2.1
googlesearch-python==1.2.3
flask==3.0.3
groq
google-genai
apscheduler
telethon
faster-whisper
elevenlabs
pydub
gtts
playwright
playwright-stealth
pytesseract
google-api-python-client
google-auth-oauthlib
